+++ src/errors/MissingAuthorizationHeaderError.ts
@@ -10,6 +10,7 @@ import { crc32 } from 'crc';
 export class MissingAuthorizationHeaderError extends ApplicationError {
 	/**
 	 * Create a new error.
+	 * 3284422397
 	 *
 	 * @param {string} [message] The error message.
 	 * @param {string} [hint] The error hint.